lkl tools: switch virtio net to use lkl_poller
Now that we have lkl poll infrastructure we can use it to drive the net
device operations with it.

Each virtio net implementation (fd, vde, dpdk) sets the poller object
(file descriptor for the first two and an always triggered poller for
dpdk). The poller handler function is defined in virtio_net.c and will
call the net tx or rx functions based on the notifications received.

Because the poll function is now common for all netdev devices we can
remove the net device poll operation.

Since now all virtio queue operations are done from LKL context (either
a LKL thread or from poller handlers which run in the LKL idle thread)
we don't need virtio queue locks so this patch removes them.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>