tools: labs: add support for generating module skeletons
Add support for generating module skeletons from full modules. This can
be done by building the skels target.

The user can select for which labs and tasks to generate the skeletons
with the LABS and TASKS variables. The user can also select at which
TODO checkpoint to generate the skeleton.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>